+++ src/core/device/index.ts
@@ -1,3 +1,0 @@
-export * from './bluetooth.service.js';
-export * from './usb.service.js';
-export * from './media.service.js';